This code here defines global variables for the package restrictedROC

This code here defines global variables for the package restrictedROC

Usage

default_part_colors

Format

An object of class character of length 2.